Fixed a GC bug, serv.time, regens, combat_swing_norange, stat bugs when their value was too high, ItemHitpointsUpdate value not being used, various wrong time conversions.
Fixed: objects not deleted correctly, then removed by the garbage collection (bug introduced in the previous commit).

- Fixed: SERV.TIME and SERV.CREATE returned the time in hundredths of second instead of in tenths of second; SERV.AGE reported a wrong value due to a wrong time conversion.
- Fixed: Chars sometimes couldn't fight after the server restart if COMBAT_PREHIT was on.
- Fixed: COMBAT_SWING_NORANGE wasn't working as intended.
- Fixed: MinCharDeleteTime ini value wasn't read correctly.
- Fixed: For damageable items (CCItemDamageable) status update, hits regen rate was used instead of the ItemHitpointsUpdate ini setting.
- Fixed: Regen rates were incorrectly converted from seconds to the internally used milliseconds, breaking them.
- Fixed: Food stat still used internally a signed short instead of an ushort data type, limiting its max value.
- Fixed: Stats and derived values still had somewhere the limit of a short type instead of using a ushort.
- Changed: Stats and maxstats mods could only have a value between 32767 and 32767, changed it to -65535 and 65535 (which is the actual new stat limit).
- Fixed: Vendors restocked after a wrong amount of time.
- Fixed: The time required to wait before another potion use was of 1,5 seconds instead of 15.

+++ src/game/CServerConfig.h
@@ -1117,9 +1117,9 @@ public:
 #define AUTOTOOLTIP_FLAG_SPELLBOOK     0x0040   // spell added to spellbook
 	uint m_iAutoTooltipResend;       // automatically resend tooltip
 
-	int64   m_iRegenRate[STAT_QTY]; // Regen's delay for each stat.
-    int     _iItemHitpointsUpdate;  // Update period for CCItemDamageable
-	int64   _iTimerCall;            // Amount of minutes (converted to ticks internally) to call f_onserver_timer (0 disables this, default).
+	int64   m_iRegenRate[STAT_QTY]; // Regen's delay for each stat (in seconds in the ini, then converted to msecs).
+    int64   _iItemHitpointsUpdate;  // Update period for CCItemDamageable (in seconds in the ini, then converted to msecs).
+	int64   _iTimerCall;            // Amount of minutes (converted to milliseconds internally) to call f_onserver_timer (0 disables this, default).
 	bool    m_bAllowLightOverride;  // Allow manual sector light override?
 	CSString m_sZeroPoint;          // Zero point for sextant coordinates counting. Comment this line out if you are not using ML-sized maps.
 	bool    m_bAllowBuySellAgent;   // Allow rapid Buy/Sell through Buy/Sell agent.
